Move app events to separate module
